test(log): move the logging-state guard into the unit suite
test_traceback_suppression_leaves_logging_state_unchanged drives a context
manager and reads logging state — it needs no database, no Prefect flow run and
none of the component suite's fixtures, so it belongs next to the other
infrahub.log tests in the unit suite.

Move the install/remove cycle it drives to tests/helpers/log.py as
traceback_suppression, so the webhook fixture and the guard share one
implementation across suites and any future fixture that installs the filter
inherits the guard. Its docstring keeps the line that says what it does; the
paragraph on why it installs the filter alone rather than calling
configure_logging is the rule the testing guidelines now carry, which the
guidelines point at the helper for.
