+++ backend/tests/component/webhook/test_traceback_suppression.py
@@ -1,14 +1,12 @@
 from __future__ import annotations
 
 import logging
-from contextlib import contextmanager
 from typing import TYPE_CHECKING
 
 import httpx
 import pytest
 from prefect import flow, task
 
-from infrahub.log import PREFECT_RUN_LOGGERS, TracebackSuppressionFilter, install_traceback_suppression_filter
 from infrahub.webhook.classifier import (
     EXPECTED_DELIVERY_ERRORS,
     ClassifiedFailure,
@@ -16,9 +14,10 @@ from infrahub.webhook.classifier import (
     WebhookDeliveryError,
     WebhookFailureClassifier,
 )
+from tests.helpers.log import traceback_suppression
 
 if TYPE_CHECKING:
-    from collections.abc import Generator, Iterator, Sequence
+    from collections.abc import Generator
 
 CLASSIFIED_MESSAGE = "The target responded with HTTP 404."
 
@@ -50,28 +49,9 @@ async def _send_classifying_in_task() -> None:
     await _classify_transport_failure()
 
 
-@contextmanager
-def _traceback_suppression() -> Iterator[TracebackSuppressionFilter]:
-    """Register the traceback filter on the Prefect run loggers, as production startup does, then remove it.
-
-    Only the filter is installed, not the whole of configure_logging: that startup routine also raises
-    the root log level, replaces the root handler and reconfigures structlog, none of which these
-    assertions need and none of which it undoes. Called per test it would leak that state into every
-    test that follows in the same worker — overriding the WARNING root level the suite pins in
-    pytest_configure, so unrelated tests drown in DEBUG records from the database driver and the HTTP
-    client.
-    """
-    traceback_filter = install_traceback_suppression_filter()
-    try:
-        yield traceback_filter
-    finally:
-        for prefect_logger_name in PREFECT_RUN_LOGGERS:
-            logging.getLogger(prefect_logger_name).removeFilter(traceback_filter)
-
-
 @pytest.fixture
 def traceback_suppression_installed() -> Generator[None, None, None]:
-    with _traceback_suppression():
+    with traceback_suppression():
         yield
 
 
@@ -121,27 +101,3 @@ async def test_unclassified_failure_logs_a_traceback(
     logged_exceptions = [record.exc_info[1] for record in caplog.records if record.exc_info]
     assert [type(exc) for exc in logged_exceptions] == [RuntimeError]
     assert str(logged_exceptions[0]) == "boom"
-
-
-def _run_logger_filters() -> dict[str, Sequence[object]]:
-    # Logger.filters is a union of filter forms; the identity of what is attached is all that matters here.
-    return {name: list(logging.getLogger(name).filters) for name in PREFECT_RUN_LOGGERS}
-
-
-def test_traceback_suppression_leaves_logging_state_unchanged() -> None:
-    """The suppression context must hand logging back exactly as it found it.
-
-    The assertions above cannot tell this apart from calling configure_logging, which installs the same
-    filter — only the process-wide state that routine also changes, and never restores, separates them.
-    """
-    root_logger = logging.getLogger()
-    level_before, filters_before = root_logger.level, _run_logger_filters()
-
-    with _traceback_suppression() as traceback_filter:
-        assert _run_logger_filters() == {
-            name: [*filters_before[name], traceback_filter] for name in PREFECT_RUN_LOGGERS
-        }
-        assert root_logger.level == level_before
-
-    assert _run_logger_filters() == filters_before
-    assert root_logger.level == level_before
